Refactor HTTP subscription tests
